// src/modules/garmentsProduction/reports/sewing/sewing-input-output-report/sewing-input-output-report-type.ts
export type SewingInputDetailsReportType = {
    SEWINGINPUTDATE: string;
    PONO?: string;
    BUYER?: string;
    STYLENO?: string;
    COLORNAME?: string;
    SIZENAME?: string;
    LINENAME?: string;
    SEWINGINPUTQTY: number;
    REMARKS?: string;
    COMPAY_NAME?: string;
    COMPANY_ADDRESS?: string;
    COMPANY_PREFIX?: string;
};


export type SewingOutputDetailsReportType = {
    SEWINGDATE: string;
    PONO?: string;
    BUYER?: string;
    STYLENO?: string;
    COLORNAME?: string;
    SIZENAME?: string;
    LINENAME?: string;
    SEWINGOUTPUT: number;
    REMARKS?: string;
    COMPAY_NAME?: string;
    COMPANY_ADDRESS?: string;
    COMPANY_PREFIX?: string;
};